Prevent propagating syntax errors from page count variables
In order to count the number of memory pages in use at the time the
oom-killer is invoked, we first have to search through the
"Mem Info" summary stored in the variables $DATASET and $MEMINFO. If the
page count values we are looking for appear more than once, then they
are stored in our variables as multiline strings instead of integers.
This leads to surprising syntax failures later when we try to perform
arithmetic with these variables:

    $ oomparser.sh
    oomparser.sh: line 97: [[: 2097038
    8388152: syntax error in expression (error token is "8388152")
    /home/ekovsky/bin/oomparser.sh: line 111: ((: 789278.00
    .38
    1 : syntax error: invalid arithmetic operator (error token is ".00
    .38
    1 ")
    oomparser.sh: line 246: [[: 2097038
    2097038: syntax error in expression (error token is "2097038")
    oomparser.sh: line 252: ((: .39
    210493078.00
    .38
    0 : syntax error: operand expected (error token is ".39
    210493078.00
    .38
    0 ")

To avoid this, we can have Awk exit immediately after the first match is
printed and stored.